refactor(integration)!: integration  object more like PHP version (#821)
BREAKING CHANGE:
In the player renderer, the customScripts property of IPlayerModel was removed, as the custom scripts should be added to the lists of JS scripts of the integration object (and IPlayermModel) to make sure that the scripts are also used in iframes created by the player client. You can remove all code that outputs the contents of IPlayerModel.customScripts.