feat(h5p-server)!: CSRF tokens in UrlGenerator (#1082)
The UrlGenerator can now generate CSRF tokens for the current user.
The editor and player correctly set user objects in the integration (needed for xAPI to work).
Closes #1012 .

BREAKING CHANGES:
The signature of H5PPlayer.render(...) was changed. You now need to pass in a user object and the optional parameters were moved to a dedicated options object.